# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Application",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/ProductForComputer": {
            "get": {
                "tags": [
                    "ProductForComputer"
                ],
                "summary": "Get a list of computer Product For Computer",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ProductForComputer"
                ],
                "summary": "Create new Product For Computer",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "description": "product_id",
                                        "type": "number"
                                    },
                                    "cat_id": {
                                        "description": "acat_id",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validate error"
                    },
                    "200": {
                        "description": "Success create element"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/ProductForComputer/category/{id}": {
            "get": {
                "tags": [
                    "ProductForComputer"
                ],
                "summary": "Get one product by category id",
                "description": "",
                "operationId": "cat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/ProductForComputer/{id}": {
            "get": {
                "tags": [
                    "ProductForComputer"
                ],
                "summary": "Get one ProductForComputer with all atributes",
                "description": "",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ProductForComputer"
                ],
                "summary": "Updates a ProductForComputer in the store with form data",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of ProductForComputer that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "description": "product_id",
                                        "type": "number"
                                    },
                                    "cat_id": {
                                        "description": "cat_id",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Error in server"
                    },
                    "400": {
                        "description": "Validate error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}